use generic host in tester app
