+++ jobs/Backend/Task/ExchangeRateUpdater/Program.cs
@@ -1,43 +1,39 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Configuration;
+using ExchangeRateUpdater.Services;
 
 namespace ExchangeRateUpdater
 {
     public static class Program
     {
-        private static IEnumerable<Currency> currencies = new[]
-        {
-            new Currency("USD"),
-            new Currency("EUR"),
-            new Currency("CZK"),
-            new Currency("JPY"),
-            new Currency("KES"),
-            new Currency("RUB"),
-            new Currency("THB"),
-            new Currency("TRY"),
-            new Currency("XYZ")
-        };
-
         public static void Main(string[] args)
         {
-            try
-            {
-                var provider = new ExchangeRateProvider();
-                var rates = provider.GetExchangeRates(currencies);
-
-                Console.WriteLine($"Successfully retrieved {rates.Count()} exchange rates:");
-                foreach (var rate in rates)
+            var host = Host.CreateDefaultBuilder(args)
+                .ConfigureAppConfiguration(
+                (context, builder) =>
                 {
-                    Console.WriteLine(rate.ToString());
-                }
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine($"Could not retrieve exchange rates: '{e.Message}'.");
-            }
+                    builder
+                        .SetBasePath(Directory.GetCurrentDirectory())
+                        .AddJsonFile("appsettings.json", optional: false);
+                })
+                .ConfigureServices(
+                (context, services) =>
+                {
+                    ConfigureServices(context.Configuration, services);
+                })
+                .Build();
 
-            Console.ReadLine();
+            host.Run();
+        }
+
+        private static void ConfigureServices(IConfiguration configuration, IServiceCollection services)
+        {
+            services.AddHostedService<ExchangeRateUpdaterJob>();
         }
     }
 }
